# swagger-spec.yaml
openapi: 3.0.0
info:
  title: API de Postagem Social
  version: 1.0.0
  description: API para postar conteúdo em diversas redes sociais.
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    SocialPostRequest:
      type: object
      properties:
        text:
          type: string
        images:
          type: array
          items:
            type: string
            format: byte
        tags:
          type: array
          items:
            type: string
paths:
...
  /threads/post:
    post:
      summary: Cria um novo post no Threads.
      tags:
        - Threads
      description: >
        Publica um novo post no Threads. A função se adapta ao conteúdo
        fornecido:

        1.  **Apenas Texto:** Cria um post de texto simples.

        2.  **Texto e 1 Imagem:** Cria um post de imagem única com legenda.

        3.  **Texto e 2+ Imagens:** Cria um post em carrossel.


        **Corpo da Requisição:**

        * **`text`** (string, opcional): O conteúdo do post, que servirá como legenda para posts com imagem/carrossel.

        * **`images`** (array, opcional): Uma lista de imagens no formato Data URL (base64). As imagens são primeiro enviadas para um serviço de hospedagem (Cloudinary) para gerar URLs públicas.

        * **`tags`** (array, opcional): Apenas a **primeira tag** da lista será usada como `topic_tag` ou adicionada como hashtag, de acordo com as regras da API do Threads para cada tipo de post.


        **Corpo da Resposta:**

        * Retorna um objeto com o ID do post criado com sucesso (status `201 Created`).

        * Retorna um erro `400 Bad Request` se os dados forem insuficientes (sem texto ou imagem).
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              allOf:
                - $ref: "#/components/schemas/SocialPostRequest"
                - type: object
                  properties:
                    text:
                      type: string
                    images:
                      type: array
                      items:
                        type: base64
                    tags:
                      type: array
                      items:
                        type: string
            example:
              text: Este é um tweet de exemplo!
              tags:
                - api
                - teste
              images:
                - data:image/png;base64,iVBORw0KGgo...
      responses:
        "201":
          description: Post criado com sucesso.
  /tumblr/blogs:
    get:
      summary: Retorna a lista de blogs do usuário no Tumblr.
      tags:
        - Tumblr
      description: >
        Recupera a lista de blogs associados à conta do Tumblr autenticada na
        API. O retorno inclui o nome de exibição (título) e o identificador
        único de cada blog. O identificador (`name`) é o valor que deve ser
        usado no parâmetro `blogName` ao criar um novo post.


        **Corpo da Resposta:**

        * Retorna um objeto contendo um array de blogs, onde cada objeto possui as chaves `name` e `title` (status `200 OK`).

        * Retorna um erro `500 Internal Server Error` se houver uma falha na comunicação com a API do Tumblr.
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Lista de blogs retornada com sucesso.
          content:
            application/json:
              schema:
                type: object
                properties:
                  blogs:
                    type: array
                    items:
                      type: object
                      properties:
                        name:
                          type: string
                          description: O identificador do blog, usado nas requisições de postagem.
                          example: meu-blog-principal
                        title:
                          type: string
                          description: O título de exibição do blog.
                          example: O Blog Incrível de Testes
  /tumblr/post:
    post:
      summary: Cria um novo post em um blog específico do Tumblr.
      tags:
        - Tumblr
      description: >
        Cria um novo post em um blog específico do Tumblr. A postagem pode ser
        de texto ou um photoset com múltiplas imagens.


        **Corpo da Requisição:**

        * **`blogName`** (string, obrigatório): O identificador do blog do Tumblr onde o post será publicado (ex: "meu-blog-principal").

        * **`text`** (string, opcional): O corpo do texto ou a legenda para o post de imagem.

        * **`images`** (array, opcional): Uma lista de imagens no formato Data URL (base64) para criar um photoset.

        * **`tags`** (array, opcional): Uma lista de tags para associar ao post.


        **Corpo da Resposta:**

        * Retorna um objeto com os dados do post criado com sucesso (status `201 Created`).

        * Retorna um erro `400 Bad Request` se o `blogName` não for fornecido ou se a API do Tumblr rejeitar a formatação.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              allOf:
                - $ref: "#/components/schemas/SocialPostRequest"
                - type: object
                  properties:
                    blogName:
                      type: string
                    text:
                      type: string
                    images:
                      type: array
                      items:
                        type: base64
                    tags:
                      type: array
                      items:
                        type: string
                  required:
                    - blogName
            example:
              blogName: meu-blog-principal
              text: Este é um tweet de exemplo!
              tags:
                - api
                - teste
              images:
                - data:image/png;base64,iVBORw0KGgo...
      responses:
        "201":
          description: Post criado com sucesso.
...